rename common files
