+++ neural_speed/core/CMakeLists.txt
@@ -53,7 +53,7 @@ function(add_test_target src)
   target_compile_options(${test_target} PRIVATE -fsanitize=address)
   target_link_options(${test_target} PRIVATE -fsanitize=address)
   target_include_directories(${test_target} PUBLIC .)
-  target_link_libraries(${test_target} PUBLIC Threads::Threads jblas::jblas ne_vec)
+  target_link_libraries(${test_target} PUBLIC Threads::Threads bestla::bestla ne_vec)
   if(NOT WIN32)
     target_link_libraries(${test_target} PUBLIC rt)
   endif()
