+++ neural_speed/models/bloom/bloom.cpp
@@ -31,7 +31,7 @@
 #include "core/data_types.h"
 #include "core/ne.h"
 #include "core/ne_layers.h"
-#include "core/ne_jblas.h"
+#include "core/ne_bestla.h"
 #include "core/layers/mha_dense.h"
 #include "models/model_utils/model_config.h"
 #include "models/model_utils/model_utils.h"
