+++ neural_speed/models/llama/llama.cpp
@@ -33,7 +33,7 @@
 #include "core/data_types.h"
 #include "core/layers/mha_dense.h"
 #include "core/ne.h"
-#include "core/ne_jblas.h"
+#include "core/ne_bestla.h"
 #include "core/ne_layers.h"
 #include "models/model_utils/model_config.h"
 #include "models/model_utils/model_files.h"
