+++ neural_speed/models/model_utils/model_files.h
@@ -29,7 +29,7 @@
 #define NE_MEM_ALIGN 16
 #endif
 
-#include "core/layers/jblas_common.hpp"
+#include "core/layers/bestla_common.hpp"
 #include "core/ne_layers.h"
 #include "models/model_utils/util.h"
 #include "models/models.h"
