+++ neural_speed/models/whisper/whisper.cpp
@@ -30,7 +30,7 @@
 #include "core/data_types.h"
 #include "core/layers/mha_dense.h"
 #include "core/ne.h"
-#include "core/ne_jblas.h"
+#include "core/ne_bestla.h"
 #include "core/ne_layers.h"
 #include "models/model_utils/model_config.h"
 #include "models/model_utils/model_utils.h"
